+++ rts/python/scalar.py
@@ -596,6 +596,10 @@ def futhark_sqrt64(x):
     return np.sqrt(x)
 
 
+def futhark_rsqrt64(x):
+    return 1 / np.sqrt(x)
+
+
 def futhark_cbrt64(x):
     return np.cbrt(x)
 
@@ -730,6 +734,10 @@ def futhark_sqrt32(x):
     return np.float32(np.sqrt(x))
 
 
+def futhark_rsqrt32(x):
+    return np.float32(1 / np.sqrt(x))
+
+
 def futhark_cbrt32(x):
     return np.float32(np.cbrt(x))
 
@@ -862,6 +870,10 @@ def futhark_log1p_16(x):
 
 def futhark_sqrt16(x):
     return np.float16(np.sqrt(x))
+
+
+def futhark_rsqrt16(x):
+    return np.float16(1 / np.sqrt(x))
 
 
 def futhark_cbrt16(x):