+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* New math functions: `f16.rsqrt`, `f32.rsqrt`, `f64.rsqrt`.
+
 ### Removed
 
 ### Changed